# Speed up tests by using fast scrypt params