--- a/core/internal/cltest/cltest.go
+++ b/core/internal/cltest/cltest.go
@@ -223,6 +223,8 @@ func NewTestConfig(t testing.TB, options ...interface{}) *TestConfig {
 	rawConfig.Set("MINIMUM_CONTRACT_PAYMENT", minimumContractPayment.Text(10))
 	rawConfig.Set("ROOT", rootdir)
 	rawConfig.Set("SESSION_TIMEOUT", "2m")
+	rawConfig.Set("INSECURE_FAST_SCRYPT", "true")
+	rawConfig.Set("BALANCE_MONITOR_ENABLED", "false")
 	rawConfig.SecretGenerator = mockSecretGenerator{}
 	config := TestConfig{t: t, Config: rawConfig}
 	return &config