--- a/core/internal/cltest/fixtures.go
+++ b/core/internal/cltest/fixtures.go
@@ -11,11 +11,6 @@ import (
 	"github.com/tidwall/gjson"
 )
 
-const (
-	// Key3cb8e3fd9d27e39a5e9e6852b0e96160061fd4ea is a valid ethereum key encrypted with the password: 'password'
-	Key3cb8e3fd9d27e39a5e9e6852b0e96160061fd4ea = `{"address":"3cb8e3fd9d27e39a5e9e6852b0e96160061fd4ea","crypto":{"cipher":"aes-128-ctr","ciphertext":"7515678239ccbeeaaaf0b103f0fba46a979bf6b2a52260015f35b9eb5fed5c17","cipherparams":{"iv":"87e5a5db334305e1e4fb8b3538ceea12"},"kdf":"scrypt","kdfparams":{"dklen":32,"n":262144,"p":1,"r":8,"salt":"d89ac837b5dcdce5690af764762fe349d8162bb0086cea2bc3a4289c47853f96"},"mac":"57a7f4ada10d3d89644f541c91f89b5bde73e15e827ee40565e2d1f88bb0ac96"},"id":"c8cb9bc7-0a51-43bd-8348-8a67fd1ec52c","version":3}`
-)
-
 // MustHelloWorldAgreement returns the fixture hello world agreement
 func MustHelloWorldAgreement(t *testing.T) string {
 	template := MustReadFile(t, "testdata/hello_world_agreement.json")