Before:

`DATABASE_URL= go test -p 4 ./... -parallel 4  798.27s user 53.55s system 446% cpu 3:10.68 total`

After:

`DATABASE_URL= go test -p 4 ./... -parallel 4  149.00s user 32.66s system 145% cpu 2:05.26 total`

Reduction of wall clock time by ~34%.
Reduction of CPU time by ~82%.