@GridTechs 2020-10-30
Thank you

On Fri, Oct 30, 2020 at 2:49 AM Sam <notifications@github.com> wrote:

> *@samsondav* commented on this pull request.
> ------------------------------
>
> In core/internal/features_test.go
> <https://github.com/smartcontractkit/chainlink/pull/3521#discussion_r514979911>
> :
>
> > @@ -118,143 +118,6 @@ func TestIntegration_HttpRequestWithHeaders(t *testing.T) {
>  	eth.EventuallyAllCalled(t)
>  }
>
> -func TestIntegration_FeeBump(t *testing.T) {
>
> This test was always broken, because the speed of decrypting keys has gone
> up it exposed the problem.
>
> I figured it was easier to remove it because it applies only to legacy TXM
> and we're removing that anyway.
>
> Instead, I fixed the test in the latest commit.
>
> —
> You are receiving this because you are subscribed to this thread.
> Reply to this email directly, view it on GitHub
> <https://github.com/smartcontractkit/chainlink/pull/3521#discussion_r514979911>,
> or unsubscribe
> <https://github.com/notifications/unsubscribe-auth/AMOWX3I74FJEHY2AZ44NDATSNKD3FANCNFSM4TDVFFUQ>
> .
>